Move comment into own component
